Optimise for speed and my own CPU
I will not share the binary itself, so having it optimise for my
generation, family, or architecture isn't in my interests. I wish for it
to be specifically targetting my CPU, no-one else's.

I also happen to not care about the file size being a whopping megabyte
or twenty larger; we're in the 21st century and can afford a terabyte if
it's required for a better experience.